Implement popa and treat add MemR7Imm7s_0, Ax_8 strictly
Before I was unsure about whether MemR7Imm7s_0 was 7 bits + sign bit or 7 bits including the sign bit. But, popa (which is from Teak Lite 2) is encoded the same way but with the 8th bit set, which implies the 8th bit is always 0 for that form of add.

Before:

    data:01c8 b1 47           add        [r7+0x31],a1

After:

    data:01c8 b1 47           popa       b1