Use preprocessor macros to handle nocash syntax differences
This at least lets everything be on one line, though $(OB) and $(CB) are not the nicest labels. Unfortunately $([) and $(]) aren't legal preprocessor macro names, so these seem to be the best option.

This allows the MemR01 tables to be more cleanly organized, hopefully reducing the risk of future typos when I do the other tables.