Remove implied fields
SLEIGH actually supports using the register directly, so there's no need for this hack. We still do need the copy hack though.

Nocash means something different with implied: operands taht aren't explicitly listed in the instruction, but still exist. That's different from r6 being in the instruction (but not corresponding to a specific bit).